Fix: Compile without warnings, fixes lots of printf format errors.
git-svn-id: svn://svn.code.sf.net/p/sipp/code/sipp/trunk@498 60913f6b-29f6-41b8-8391-1ad2d2d6c344

+++ deadcall.hpp
@@ -17,6 +17,6 @@ public:
   virtual void dump();
 
 protected:
-  int expiration;
+  unsigned long expiration;
   char *reason;
 };
